+++ contracts/mocks/docs/utils/cryptography/ERC7739SignerECDSA.sol
@@ -1,16 +1,16 @@
-// contracts/ERC7739SignerECDSA.sol
+// contracts/ERC7739ECDSA.sol
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.20;
 
 import {ECDSA} from "@openzeppelin/contracts/utils/cryptography/ECDSA.sol";
 import {EIP712} from "@openzeppelin/contracts/utils/cryptography/EIP712.sol";
 
-import {ERC7739Signer} from "../../../../utils/cryptography/ERC7739Signer.sol";
+import {ERC7739} from "../../../../utils/cryptography/ERC7739.sol";
 
-contract ERC7739SignerECDSA is ERC7739Signer {
+contract ERC7739ECDSA is ERC7739 {
     address private immutable _signer;
 
-    constructor(address signerAddr) EIP712("ERC7739SignerECDSA", "1") {
+    constructor(address signerAddr) EIP712("ERC7739ECDSA", "1") {
         _signer = signerAddr;
     }