+++ contracts/mocks/utils/cryptography/ERC7739SignerRSAMock.sol
@@ -3,11 +3,11 @@
 pragma solidity ^0.8.20;
 
 import {EIP712} from "@openzeppelin/contracts/utils/cryptography/EIP712.sol";
-import {ERC7739Signer} from "../../../utils/cryptography/ERC7739Signer.sol";
+import {ERC7739} from "../../../utils/cryptography/ERC7739.sol";
 import {SignerRSA} from "../../../utils/cryptography/SignerRSA.sol";
 
-contract ERC7739SignerRSAMock is ERC7739Signer, SignerRSA {
-    constructor(bytes memory e, bytes memory n) EIP712("ERC7739SignerRSA", "1") {
+contract ERC7739RSAMock is ERC7739, SignerRSA {
+    constructor(bytes memory e, bytes memory n) EIP712("ERC7739RSA", "1") {
         _initializeSigner(e, n);
     }
 }